#26 Refactor code and design
