+++ src/main/java/com/andersen/corgisteam/corgiroulette/dto/ResponseGroupedMarksDto.java
@@ -6,9 +6,6 @@ public class ResponseGroupedMarksDto {
     private User user;
 
     private double mark;
-
-    public ResponseGroupedMarksDto() {
-    }
     
     public User getUser() {
         return user;
