+++ src/main/java/com/andersen/corgisteam/corgiroulette/service/PairService.java
@@ -1,10 +1,11 @@
 package com.andersen.corgisteam.corgiroulette.service;
 
 import com.andersen.corgisteam.corgiroulette.entity.Pair;
+import com.andersen.corgisteam.corgiroulette.entity.User;
 
 public interface PairService {
 
     Pair getPair();
 
-    Pair changeOpponent(Pair pair);
+    Pair changeOpponent(User userToSave, User userToChange);
 }
