+++ src/main/resources/db.properties
@@ -1,4 +1,4 @@
-jdbcUrl=jdbc:postgresql://35.228.230.177:5432/postgres
+jdbcUrl=jdbc:postgresql://localhost:5432/corgiroulettedb
 driverClassName=org.postgresql.Driver
 dataSource.user=postgres
 dataSource.password=postgres
