+++ src/main/resources/db/migration/V5__Create_marks_table.sql
@@ -1,10 +1,10 @@
 CREATE TABLE IF NOT EXISTS marks
 (
-    id          BIGSERIAL NOT NULL,
-    user_id     BIGSERIAL NOT NULL,
-    mark        DECIMAL(4,1) NOT NULL,
-    date_time   TIMESTAMP NOT NULL,
+    id        BIGSERIAL     NOT NULL,
+    user_id   BIGSERIAL     NOT NULL,
+    mark      DECIMAL(4, 1) NOT NULL,
+    date_time TIMESTAMP     NOT NULL,
     constraint pk_mark primary key (id),
     CONSTRAINT fk_user_mark FOREIGN KEY (user_id) REFERENCES users (id)
         ON DELETE CASCADE
-    );
+);
