+++ src/test/java/com/andersen/corgisteam/corgiroulette/service/FindOpponentsUsingListTest.java
@@ -5,7 +5,10 @@ import com.andersen.corgisteam.corgiroulette.entity.Pair;
 import com.andersen.corgisteam.corgiroulette.entity.Team;
 import com.andersen.corgisteam.corgiroulette.entity.User;
 import com.andersen.corgisteam.corgiroulette.repository.*;
+import com.andersen.corgisteam.corgiroulette.repository.impl.TeamRepositoryImpl;
+import com.andersen.corgisteam.corgiroulette.repository.impl.UserRepositoryImpl;
 import com.andersen.corgisteam.corgiroulette.service.exception.NullListForGeneratePairException;
+import com.andersen.corgisteam.corgiroulette.service.impl.FindOpponentsUsingList;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
